React Reorganize - upper level
